[chore] rename penSetting and penCard
