+++ src/components/Editor/PenSettingModal.vue
